third_party/WebKit/LayoutTests/media/track/track-webvtt-tc010-no-timings.html

<!DOCTYPE html>
<title>Tests cue without timings are ignored.</title>
<script src="track-helpers.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<video>
    <track src="captions-webvtt/tc010-no-timings.vtt" default>
</video>
<script>
async_test(function(t) {
    var track = document.querySelector("track");

    track.onload = t.step_func_done(function() {
        assert_cues_equal(track.track.cues, []);
    });
});
</script>
